quran_bot_full/requirements.txt

discord.py[voice]>=2.5
yt-dlp>=2025.4
imageio-ffmpeg
mutagen
aiohttp
apscheduler
